Add basic detection for some psql commands
This detects the `\d <table_name>` and `\dt` metacommands, and
suggests an alternative to use instead. The metacommands all make
calls to `pg_catalog` tables/views, which Logfire (backed by
Datafusion) dosn't support, and some run a bunch of queries, so
it's not really easy to mock them out, and better to just show an
error and show the user the right query to run instead. More
metacommand detection will be added in the future, but this fine
for now I think as an initial attempt.

Signed-off-by: Matthew Peveler <mpeveler@tigerdata.com>